services/gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
grpcio==1.59.3
grpcio-tools==1.59.3
protobuf==4.25.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.25.2
redis==5.0.1
nats-py==2.3.1
prometheus-client==0.19.0
structlog==23.2.0
tenacity==8.2.3
email-validator==2.1.0
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
